+++ CHANGELOG.md
@@ -24,7 +24,7 @@ each written as `field_name:` followed by its value.
 
  **V 1.8.5** &ensp; <@leonardo-bersellini> &emsp; 07 . 09 . 2026
 
- `commit:` Stack templates *-o- ()*
+ `commit:` Stack templates *-o- (13fdcac)*
  <br>
  `scope:` Utils, Codegen, Semantics
  <br>
@@ -34,7 +34,7 @@ each written as `field_name:` followed by its value.
 
  **V 1.8.4** &ensp; <@leonardo-bersellini> &emsp; 06 . 09 . 2026
 
- `commit:` Removed Strings *-o- (cef441d)*
+ `commit:` Removed Strings *-o- (fa99422)*
  <br>
  `scope:` Types, Semantics
  <br>