+++ CHANGELOG.md
@@ -22,6 +22,26 @@ each written as `field_name:` followed by its value.
 
  ---
 
+ **V 1.8.7** &ensp; <@leonardo-bersellini> &emsp; 10 . 09 . 2026
+
+ `commit:` Refactored Symbols Structure *-o- ()*
+ <br>
+ `scope:` Symbols, Semantics
+ <br>
+ `features:` refactored symbols structure and tables
+
+ ---
+
+ **V 1.8.6** &ensp; <@leonardo-bersellini> &emsp; 09 . 09 . 2026
+
+ `commit:` Added Integration Test *-o- (12c7eeb)*
+ <br>
+ `scope:` Tests
+ <br>
+ `features:` new tests that run a whole bismuth source code
+
+ ---
+
  **V 1.8.5** &ensp; <@leonardo-bersellini> &emsp; 07 . 09 . 2026
 
  `commit:` Stack templates *-o- (13fdcac)*